Separate ssl certs for PROD and INTEGRATION tests. First working integration test added.
