+++ src/integration/groovy/pl/astronomy/arqonia20/api/user/SignupEndpointTest.groovy
@@ -1,6 +1,5 @@
 package pl.astronomy.arqonia20.api.user
 
-
 import org.springframework.http.HttpMethod
 import pl.astronomy.arqonia20.BaseIntegrationTest
 import pl.astronomy.arqonia20.domain.user.User
@@ -8,17 +7,6 @@ import pl.astronomy.arqonia20.domain.user.UserRole
 import pl.astronomy.arqonia20.domain.user.UserRoleType
 
 class SignupEndpointTest extends BaseIntegrationTest {
-
-    def "test should pass!"() {
-        given:
-        def var1 = "correctValue"
-
-        when:
-        def response = "correctValue"
-
-        then:
-        response == var1
-    }
 
     def "should register new normal user"() {
         given:
@@ -30,9 +18,7 @@ class SignupEndpointTest extends BaseIntegrationTest {
         )
 
         when:
-        // TODO Implement 'password' flow for integration tests
-        // URL: https://www.baeldung.com/how-to-use-resttemplate-with-basic-authentication-in-spring
-        def response = restTemplate().exchange(
+        def response = restTemplate.exchange(
                 localUrl("/signup"),
                 HttpMethod.POST,
                 prepareEntity(userDto),
@@ -41,6 +27,5 @@ class SignupEndpointTest extends BaseIntegrationTest {
         then:
         response.statusCode.value() == 201
     }
-
 
 }
