+++ src/main/resources/application-integration.yml
@@ -4,3 +4,13 @@ spring:
       database: arqonia2
       port: 27017
       uri: mongodb://localhost
+
+server:
+  ssl:
+    enabled: true # set to FALSE, if want to disable HTTPS and switch back to HTTP
+    key-store-type: JKS
+    key-store: classpath:keystore/https.integration.tests/arqonia_https_it.jks
+    key-store-password: pass123
+    key-alias: arqonia_https_it
+    trust-store: classpath:keystore/https.integration.tests/arqonia_https_it.jks
+    trust-store-password: pass123
