+++ src/main/resources/application.yml
@@ -11,10 +11,10 @@ server:
   ssl:
     enabled: true # set to FALSE, if want to disable HTTPS and switch back to HTTP
     key-store-type: JKS
-    key-store: classpath:keystore/https/arqonia_https.jks
+    key-store: classpath:keystore/https.prod/arqonia_https.jks
     key-store-password: pass123
     key-alias: arqonia_https
-    trust-store: classpath:keystore/https/arqonia_https.jks
+    trust-store: classpath:keystore/https.prod/arqonia_https.jks
     trust-store-password: pass123
 
 #management:
