+++ src/main/resources/keystore/https.integration.tests/ca/arqonia_ca_certificate_it.pem
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDrTCCApWgAwIBAgIJAM93QhPZ9+nUMA0GCSqGSIb3DQEBCwUAMG0xCzAJBgNV
+BAYTAlBMMRUwEwYDVQQIDAxXaWVsa29wb2xza2ExDzANBgNVBAcMBlBvem5hbjEQ
+MA4GA1UECgwHQXJxb25pYTEQMA4GA1UECwwHQXJxb25pYTESMBAGA1UEAwwJbG9j
+YWxob3N0MB4XDTE5MDEyNDIxMzgzMloXDTI5MDEyMTIxMzgzMlowbTELMAkGA1UE
+BhMCUEwxFTATBgNVBAgMDFdpZWxrb3BvbHNrYTEPMA0GA1UEBwwGUG96bmFuMRAw
+DgYDVQQKDAdBcnFvbmlhMRAwDgYDVQQLDAdBcnFvbmlhMRIwEAYDVQQDDAlsb2Nh
+bGhvc3QwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDT8MCyCg7hI33k
+OaiSM+9NDys9qlFrNdAc6BX9a4BZCEGJPDSDngLOuQMR5Tfwat7z729nZ6WIu8c9
+Okfg5yXXvthA0OQKyOv4HPNBm+bCHKTYNs10/H15LiTbLdgYUqxTsMIVCRWaTQuQ
+OBbZkE8Wv7BXkt8x+lUpAeDuxVPPyMG0nZ6c65F8Ft4pm46VBDsI34uQgtuDkMXh
+Dgmq1YBW+NnuY76JmXhJi11JV6B4C6SREqrpRUyfkLAFzUwPyR8va3vUshpWC5ml
+Ioymp2l3JNHD+YQtOzh3uh6A/Jka0w8rugexMS5bjPX/5YW7YKhoogyshrQXgShs
+4l6f+1HTAgMBAAGjUDBOMB0GA1UdDgQWBBTc4ozc3Cq15CshXXqb7Q8XRkQsAjAf
+BgNVHSMEGDAWgBTc4ozc3Cq15CshXXqb7Q8XRkQsAjAMBgNVHRMEBTADAQH/MA0G
+CSqGSIb3DQEBCwUAA4IBAQCmv9PcdcuFK/tbzBohMKPvC+ukbB6hM736THkQGEDa
+O6yJ+bNOBvfEDCPJ+xUhyw77lHKofL65IzvdBKdit/NZL3+JM4nPb2xKGiOiSBkl
+I3ycD0YZgJoCBZexaH9rmnyV81bAIwriZ0mAjJxDVCLQJj9otk+0Bb+GHBDHFejN
+AoYu4k6eHFqoQaXypmB+4XOYlqk4Qje6EyQLUdy9saRl1zoAs6GePiZaEAYPyjZV
+1fskNHz1s2KoMmnsieYFnfmg3+HPiUAWp1ZHF83mNJ8AD04f2ZRqZUGU/Abn4JsX
+U6/qs+DFfGUPxrntFdZa73rogBOvcNKSQ0EIZYDl5i59
+-----END CERTIFICATE-----
