+++ src/main/resources/keystore/https.integration.tests/ca/arqonia_https_certificate_it.cer
@@ -1,0 +1,32 @@
+-----BEGIN CERTIFICATE-----
+MIIFdjCCBF4CCQDIRqu1sc0+lDANBgkqhkiG9w0BAQUFADBtMQswCQYDVQQGEwJQ
+TDEVMBMGA1UECAwMV2llbGtvcG9sc2thMQ8wDQYDVQQHDAZQb3puYW4xEDAOBgNV
+BAoMB0FycW9uaWExEDAOBgNVBAsMB0FycW9uaWExEjAQBgNVBAMMCWxvY2FsaG9z
+dDAeFw0xOTAxMjQyMTM4MzJaFw0yOTAxMjEyMTM4MzJaMG0xCzAJBgNVBAYTAlBM
+MRUwEwYDVQQIEwxXaWVsa29wb2xza2ExDzANBgNVBAcTBlBvem5hbjEQMA4GA1UE
+ChMHQXJxb25pYTEQMA4GA1UECxMHQXJxb25pYTESMBAGA1UEAxMJbG9jYWxob3N0
+MIIDQjCCAjUGByqGSM44BAEwggIoAoIBAQCPeTXZuarpv6vtiHrPSVG28y7Fnjuv
+Nxjo6sSWHz79NgbnQ1GpxBgzObgJ58KuHFObp0dbhdARrbi0eYd1SYRpXKwOjxSz
+Nggooi/6JxEKPWKpk0U0CaD+aWxGWPhL3SCBnDcJoBBXsZWtzQAjPbpUhLYpH51k
+jviDRIZ3l5zsBLQ0pqwudemYXeI9sCkvwRGMn/qdgYHnM423krcw17njSVkvaAmY
+chU5Feo9a4tGU8YzRY+AOzKkwuDycpAlbk4/ijsIOKHEUOThjBopo33fXqFD3ktm
+/wSQPtXPFiPhWNSHxgjpfyEc2B3KI8tuOAdl+CLjQr5ITAV2OTlgHNZnAh0AuvaW
+poV499/e5/pnyXfHhe8ysjO65YDAvNVpXQKCAQAWplxYIEhQcE51AqOXVwQNNNo6
+NHjBVNTkpcAtJC7gT5bmHkvQkEq9rI837rHgnzGC0jyQQ8tkL4gAQWDt+coJsyB2
+p5wypifyRz6Rh5uixOdEvSCBVEy1W4AsNo0fqD7UielOD6BojjJCilx4xHjGjQUn
+txyaOrsLC+EsRGiWOefTznTbEBplqiuH9kxoJts+xy9LVZmDS7TtsC98kOmkltOl
+XVNb6/xF1PYZ9j897buHOSXC8iTgdzEpbaiH7B5HSPh++1/et1SEMWsiMt7lU92v
+AhErDR8C2jCXMiT+J67ai51LKSLZuovjntnhA6Y8UoELxoi34u1DFuHvF9veA4IB
+BQACggEADoGNjdLJuGdezj4dsAT8SKefkPdTqywfnRHwiFoXcd6AZR/lG+aaYY5h
+Fkbl3SFIfKkWsNdQC4jXXjpR9jhVxw0qGzCoVbwD7G2T2S2uMJVSoiIaXpfEtRRT
++c01tgLt5nnzZ6zBQFMSB7yiN/Cjlthk1p62sNd+DYP0Dhc1vkyMdCobIoo+VLMQ
+eQRmiDco5y5Hv/cj8lFQ/C4lQeP8B9dm+LdwgUcBURFKK9n5uiqPdZbxlcYI6w8T
+tL+g+7qaz8wFmVYBFhIBe2GBLrPD4UihGnccslYWwrXANUzBdEt14+ewOs6WNLLm
+tM5VbFPgFpGRFChjFxAt6G4r6MoC/TANBgkqhkiG9w0BAQUFAAOCAQEAsIpFD3U3
+c3bTnxviptidmYJoBU0tH0ajwbBGUG8EJntOdkXf3U7F7ZDaHJatnACn3NIgtHcO
+ScSfYtDqU2vNzpfdKugOXnvBW5Ahh5nJYVKuT00de0CvOnNmr5optNz5uCVHU5ON
+fWfx2bgtH7gkxQtqzGPaQ6iQTLi1FaXphOBaGegzbisRkVPmE0IDt0VIWTE+RiwM
+46Rv85tjL7jG9Xz1NkK8qEGjNo9wDR/UcQjViwalS2nxbVsVZ+xmFjVh8x8kdgZ5
+ShVf4n9sTlye8URYjx9ejAWLKITzxbyTyF7CwRDGKjMkNqXfzzS7WY+AI+0M7/yp
+s4goDjkllKIYhg==
+-----END CERTIFICATE-----
