+++ src/main/resources/keystore/https.integration.tests/ca/arqonia_https_it.csr
@@ -1,0 +1,25 @@
+-----BEGIN NEW CERTIFICATE REQUEST-----
+MIIEPzCCA+oCAQAwbTELMAkGA1UEBhMCUEwxFTATBgNVBAgTDFdpZWxrb3BvbHNr
+YTEPMA0GA1UEBxMGUG96bmFuMRAwDgYDVQQKEwdBcnFvbmlhMRAwDgYDVQQLEwdB
+cnFvbmlhMRIwEAYDVQQDEwlsb2NhbGhvc3QwggNCMIICNQYHKoZIzjgEATCCAigC
+ggEBAI95Ndm5qum/q+2Ies9JUbbzLsWeO683GOjqxJYfPv02BudDUanEGDM5uAnn
+wq4cU5unR1uF0BGtuLR5h3VJhGlcrA6PFLM2CCiiL/onEQo9YqmTRTQJoP5pbEZY
++EvdIIGcNwmgEFexla3NACM9ulSEtikfnWSO+INEhneXnOwEtDSmrC516Zhd4j2w
+KS/BEYyf+p2BgeczjbeStzDXueNJWS9oCZhyFTkV6j1ri0ZTxjNFj4A7MqTC4PJy
+kCVuTj+KOwg4ocRQ5OGMGimjfd9eoUPeS2b/BJA+1c8WI+FY1IfGCOl/IRzYHcoj
+y244B2X4IuNCvkhMBXY5OWAc1mcCHQC69pamhXj3397n+mfJd8eF7zKyM7rlgMC8
+1WldAoIBABamXFggSFBwTnUCo5dXBA002jo0eMFU1OSlwC0kLuBPluYeS9CQSr2s
+jzfuseCfMYLSPJBDy2QviABBYO35ygmzIHannDKmJ/JHPpGHm6LE50S9IIFUTLVb
+gCw2jR+oPtSJ6U4PoGiOMkKKXHjEeMaNBSe3HJo6uwsL4SxEaJY559POdNsQGmWq
+K4f2TGgm2z7HL0tVmYNLtO2wL3yQ6aSW06VdU1vr/EXU9hn2Pz3tu4c5JcLyJOB3
+MSltqIfsHkdI+H77X963VIQxayIy3uVT3a8CESsNHwLaMJcyJP4nrtqLnUspItm6
+i+Oe2eEDpjxSgQvGiLfi7UMW4e8X294DggEFAAKCAQAOgY2N0sm4Z17OPh2wBPxI
+p5+Q91OrLB+dEfCIWhdx3oBlH+Ub5pphjmEWRuXdIUh8qRaw11ALiNdeOlH2OFXH
+DSobMKhVvAPsbZPZLa4wlVKiIhpel8S1FFP5zTW2Au3mefNnrMFAUxIHvKI38KOW
+2GTWnraw134Ng/QOFzW+TIx0Khsiij5UsxB5BGaINyjnLke/9yPyUVD8LiVB4/wH
+12b4t3CBRwFREUor2fm6Ko91lvGVxgjrDxO0v6D7uprPzAWZVgEWEgF7YYEus8Ph
+SKEadxyyVhbCtcA1TMF0S3Xj57A6zpY0sua0zlVsU+AWkZEUKGMXEC3obivoygL9
+oDAwLgYJKoZIhvcNAQkOMSEwHzAdBgNVHQ4EFgQU4lj/1S3VczZ7eD9TWqslaKaP
+ZP4wDQYJYIZIAWUDBAMCBQADQAAwPQIcYdWdm0Y0sDAmCw4fBsC74MbyNynHxaxe
+McsDIAIdAKh9GTXIOMBKLiPcAWhMWPTwAxLdbfXEeZ7avS8=
+-----END NEW CERTIFICATE REQUEST-----
