+++ src/main/resources/keystore/https.integration.tests/ca/private/arqonia_ca_key_it.pem
@@ -1,0 +1,30 @@
+-----BEGIN ENCRYPTED PRIVATE KEY-----
+MIIFHzBJBgkqhkiG9w0BBQ0wPDAbBgkqhkiG9w0BBQwwDgQISLAogiouSC4CAggA
+MB0GCWCGSAFlAwQBKgQQujTV8gUxkQ9kSQoAhVmVuwSCBND/+Q1m1b8JYyGfJhuL
+fS9dMVU0kQzktSTM25Cq11I2AvSjrAb0V7oOW9uPRsKZ/+sHCnef7AA3flJomXlg
+pLZeXr11mrT+IpDOfIJnwjs98+RERscVLLoeLIpZPah2Hh7v5WmmfCfVXZvHIiy1
+YCAyYORp8g/bHgxj411FDOiqrcOwsNgIM9b46PMsgeev20UCq9A8OJv6tP229PR1
+eGsh+awb2lLWSb2l2dWRJQ2CIz+mihKpomJL0Uz8O6SR79cJtDJOCbKkEx2H7QcM
+//FBSjIxen1wSSnpujSGFuhhDiXUF9vBB8Q7VuS3AoEtlYopHQ1cr7/hzzaDX7b1
+zm3NPYzr5ExgL6ZABJ87LxxHyRn5NMIijB3GoQXwtbGyay9HUF1iGlfIAUzFbBhv
+5PU6JzT0/zCwqlf18kK71AtdiwCOVeWNFYndR47I/4h+Hn5rfEjiQxpBhvef+oGe
+P0t5ruhB+lLSkDUk/NzvFnbjHVYxJ4nxsd9xgZciCrMvNA1N/1+1YALuuBmVIENZ
+d+/iqz2d2plZ2OjVO0RH5XmW/GJD226+XELCfmYP3OdBeXelzKGO0OrAOwCpGfF3
+emT2uXEJWNh4+CW6wSDeuXd9Xx5YzDJZJpeosVAkTh0GyCnJqdlUCUGsbaiXnPJE
+sh/MLAd+f/cd3UeEhrwTptNa+y/fqzcXrSMnKUWTo4PxybEawlhGvIic9x+oWL2+
+9oNrts2M6H4vT9dmcKTuWkxH/qs4R+/0ou1soZQGkus+n1p+ZwTLIFRCCxFIhGev
+BmlGiGvN3C8YPDpyIRKtkm4LzOjbDlPzow4uVa03WUsTyTKpL0fmvMbUtS1lmMPK
+8B1BQmuEGp9jksjCGNKO5GQYSLMmWyObOssGhDBuOR2IXLCRa0JqkmROyoiRISUc
+OHOevdwneDmI7PaWUfUollh4f7ihB2qW5vC4kwhqRe/D7Fd3uNm8+QtOvhZyrf5r
++c16MsUjqnD5M1RXN+Y61LCVCD/SHCRNatvJYOKvKCkVQMTKNe1hWwer10Dj5fcV
+eWYUWT+h6rEyDrHMeghiJNG7+AS9BV7s8AhpNEvLCoIBsQfgO4CzrZOX1UCLRfg9
+N56GoqB4yVc+l7/XtRrLq4kfB5geXSErUWOiKDf2a5nqcKQgWAXEdNvieiZQLvF+
+YBs2MIhQhD3ywp+1OPmltNYtH5BH43KA8F6DwS/Y0Ok91hjIXmVFDSwxmH9EGIQO
+aTLXhVTAwzkNw8xZLqiaicd9ypppyqW+8QnGftzRBxg1GuBKqamgc428Tc0VoS3I
+Si9KpXrSQkW8rfp3PIOEioZIAQm1u2w49UozBPGeEToBN6FuUjdDbOnZqKAvZ0OT
+Ct9VW83KFd+IyruJkTeSNw0F6H9BOGyyPpmgnurOdRPt04prK6AQULqlpIyue0a0
+LjQkPS0rHq7onsUhjcpls85i/dDPnXMhxhaUyBw7yWQr0PGAN1RP0maBq42WExs2
+QIIbFA63lzLG0RN0r3orK4IVwCApwWbsPOeL2dVYOwNdkHcbU6+ngSnleASc5p0K
+l8qOpI1AOQhCKmt8zjnvWAgkDbM6vjYgXb6RF+u3gD9C6GNgSkvnqtzlnHPuu51h
+V2AI2FPx76woe5CBHHB/JKJQ3Q==
+-----END ENCRYPTED PRIVATE KEY-----
