+++ src/main/resources/keystore/https.prod/ca/arqonia_ca_certificate.pem
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDrzCCApegAwIBAgIJALC0WmCTbdtmMA0GCSqGSIb3DQEBCwUAMG4xCzAJBgNV
+BAYTAlBMMRUwEwYDVQQIDAxXaWVsa29wb2xza2ExDzANBgNVBAcMBlBvem5hbjEQ
+MA4GA1UECgwHQXJxb25pYTEQMA4GA1UECwwHQXJxb25pYTETMBEGA1UEAwwKYXJx
+b25pYS5wbDAeFw0xOTAxMjQyMTI3NTRaFw0yOTAxMjEyMTI3NTRaMG4xCzAJBgNV
+BAYTAlBMMRUwEwYDVQQIDAxXaWVsa29wb2xza2ExDzANBgNVBAcMBlBvem5hbjEQ
+MA4GA1UECgwHQXJxb25pYTEQMA4GA1UECwwHQXJxb25pYTETMBEGA1UEAwwKYXJx
+b25pYS5wbDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBALWQJ6xkoRCV
++ME+SV86qKLrhHyZkccDXbWAq+oT7Tj/+GfgVcx0s8aA0WTalUA56qH5p5iUdc3L
+Ae0DswYI33zqrZLBdLWWeT0GOHGrTBa3J9rQAMMbzewZtpLwgGwBWzaCmS1LkNE/
+MT+WS2r7K4h8W2GZ99hAapBZKiYiDoRYt0vbiP4FAq+iZg5cEi7cqOTGSmJAtHAr
+5UfJ/K+jS/nBz5OjDjtj871SA9wAfEu/2vQ+EGvDUafFHZzrEIQpqKIRhh21UCsJ
+XObtFpu3KCZCB8JGO9tSUzs3aWJfQV8GRyAYhSBYBXh+9+c63Y4OFmtkKUp899DG
+6k/PZFAxlxsCAwEAAaNQME4wHQYDVR0OBBYEFLyseWA9D5eCkThCD1ET52S5bO/K
+MB8GA1UdIwQYMBaAFLyseWA9D5eCkThCD1ET52S5bO/KMAwGA1UdEwQFMAMBAf8w
+DQYJKoZIhvcNAQELBQADggEBAJ1GCe8V8yllysfDroeoyjKush8/DrnAbKUV1rPy
+BOtRojZu+2jJTRogqkG1wDIgz6BUafJnRaNTys6lLuXECDgKGsYc1gecMA8g1iGN
+0v/K22TivrGWuU6vet7fVOP6sh+0+p3hYuPmbKyeJwEKg1MvvhgGdYEplnZMsO8p
+HS2nwoU8abcDZVx0LCELgM8p0CsYpEOjWfxMznkQdLHbu4yABZFQhRBmiwvHUB0G
+Gaq89HHOZGG/sZMgyzDCLPgOb3IN34nmPWe5bI1NzQI8qxJE212e0ZZyCmck+t7g
+8LiU/MZEukJmdhzIzIywYNQOuPvRhdIkZrWBLHcTivHXhLo=
+-----END CERTIFICATE-----
