+++ src/main/resources/keystore/https.prod/ca/arqonia_ca_certificate.srl
@@ -1,0 +1,1 @@
+A298F0F4D641474A
