+++ src/main/resources/keystore/https.prod/ca/arqonia_https_certificate.cer
@@ -1,11 +1,11 @@
 -----BEGIN CERTIFICATE-----
-MIIFeTCCBGECCQCNn0p/F99zUDANBgkqhkiG9w0BAQUFADBuMQswCQYDVQQGEwJQ
+MIIFeDCCBGACCQCimPD01kFHSjANBgkqhkiG9w0BAQUFADBuMQswCQYDVQQGEwJQ
 TDEVMBMGA1UECAwMV2llbGtvcG9sc2thMQ8wDQYDVQQHDAZQb3puYW4xEDAOBgNV
 BAoMB0FycW9uaWExEDAOBgNVBAsMB0FycW9uaWExEzARBgNVBAMMCmFycW9uaWEu
-cGwwHhcNMTkwMTE3MDk0MzE1WhcNMjkwMTE0MDk0MzE1WjBuMQswCQYDVQQGEwJQ
+cGwwHhcNMTkwMTI0MjEyNzU0WhcNMjkwMTIxMjEyNzU0WjBuMQswCQYDVQQGEwJQ
 TDEVMBMGA1UECBMMV2llbGtvcG9sc2thMQ8wDQYDVQQHEwZQb3puYW4xEDAOBgNV
 BAoTB0FycW9uaWExEDAOBgNVBAsTB0FycW9uaWExEzARBgNVBAMTCmFycW9uaWEu
-cGwwggNDMIICNQYHKoZIzjgEATCCAigCggEBAI95Ndm5qum/q+2Ies9JUbbzLsWe
+cGwwggNCMIICNQYHKoZIzjgEATCCAigCggEBAI95Ndm5qum/q+2Ies9JUbbzLsWe
 O683GOjqxJYfPv02BudDUanEGDM5uAnnwq4cU5unR1uF0BGtuLR5h3VJhGlcrA6P
 FLM2CCiiL/onEQo9YqmTRTQJoP5pbEZY+EvdIIGcNwmgEFexla3NACM9ulSEtikf
 nWSO+INEhneXnOwEtDSmrC516Zhd4j2wKS/BEYyf+p2BgeczjbeStzDXueNJWS9o
@@ -17,16 +17,16 @@ IHannDKmJ/JHPpGHm6LE50S9IIFUTLVbgCw2jR+oPtSJ6U4PoGiOMkKKXHjEeMaN
 BSe3HJo6uwsL4SxEaJY559POdNsQGmWqK4f2TGgm2z7HL0tVmYNLtO2wL3yQ6aSW
 06VdU1vr/EXU9hn2Pz3tu4c5JcLyJOB3MSltqIfsHkdI+H77X963VIQxayIy3uVT
 3a8CESsNHwLaMJcyJP4nrtqLnUspItm6i+Oe2eEDpjxSgQvGiLfi7UMW4e8X294D
-ggEGAAKCAQEAhJi+8P3rsy2UIa0UGoMy71MfaRM7TZ/KYkqkdZ0bvoQSgTJAFRkg
-Ob1vLbNjst+BXE+2Bbnj1BMMv3oSZoE8LFVsTc2xyvBpuvqyu5iJ57fpaAVXScj6
-DFAAtaEMqZOA+csR96FI693xuQ80CfZV52dPGF5jLWGKlX2/vM7WwHn4l8to972g
-DoXPTC7AjDPuoUJgKq/68Y4w76PSRmTC6T84ek9Lu5PQABwW5wRcexfKvvg/2zGC
-DRI20KWJr1jiwdowDMKGUkjgnqstBQ5bVHbHQQhpD1m4e5m7+dwzWOdy+Fo1T/n1
-BfboSoPb2GNl8/H8IbiGO7491whK3nlGjzANBgkqhkiG9w0BAQUFAAOCAQEAIW6f
-slay7QYAMqSoX01ctUSKLg5nZH186rn/ADzyFeFgIAgE9TwVp246k17zLkzjFvHu
-Zo5dDoEsg8T2f2TOMnWMhAU950IFf+b4Hya6uE9jheliC0nFcne4Z1rLn2gSyBIJ
-EEX+0pMG22UhbyOuiDNhW9uoC3PIV3KU89miIrVxphMDjMZPBAnEBdF8DJvBWOtt
-oHo6yRJFuDju6vBedJyC3LvR/HeSNaapSrmksyWp5JQ1vpkqhBAnumAHJaKZIE7t
-yK63p6Emf2QhkpHC+p7EaD9SaAHQoAu8JoUQ7e3koGVjSFnR7cwqwixwP2KELEb8
-y6ULK+5go1ISusyobg==
+ggEFAAKCAQAhCtQbM1KN9PCtJe4gw3KLaQlGOo+gk7MJBq+hDT1zpL87WB+qnjkh
+vZJkv7B7AuMvC/BXJgTJ+yGvUa2NqyfxjEktP3R4vANG8PVBBA5ddknqD8EgUwKD
+wSC5GPjZU5w6hazUZIpOHIRcZogiUw2kh6JReysI1DMDG/T6uZUcOEiM58VZwmXj
+PmbLlDnEh31Byx6sJqK7ZrsaUvcvlnWd1kEYIP6aMCFz3NrXlrpFsia7pmOb26ks
+67BGPz5pU5UUGzYWqXm0Ljh4H6XUs14JuZ8QV/+/+TBJTjD5P/eAPI1+35gQn7t8
+8dIVf20PnjdGQChUX/PgqzcFpY5idcNQMA0GCSqGSIb3DQEBBQUAA4IBAQA+7lIr
+M0EXoJetlVdrATFrdnA4tGfQIVsoHWQCusaq39bTcDauhBQmFQ16Feo53p41hQUB
+c1SEKJdwHQ3GnCe8pEi3nw7wf/E6ILeXwYotvLT39a51ouocZ4OJTdcfz10naQmH
+C1EVUXkvtuIy/a7MDiOdVdop1wywtqsULzYjcUrB/yRPoJv7Ks6QzO3r8CibfXqK
+wAZUGDkUlcTseEb575s95J9xlaEWyPvd9nN3nnLpWQPT4CrOzIvb4ZfAB6FbiH+e
+Rr/cjdBfectzBv0Rzjyn1SVsduS9SHWv4qyNbV9xVLISVd5LbySOVYykh7+z6Fxh
+G9vRgVge4DAi7YVI
 -----END CERTIFICATE-----
