+++ src/main/resources/keystore/https.prod/ca/private/arqonia_ca_key.pem
@@ -1,0 +1,30 @@
+-----BEGIN ENCRYPTED PRIVATE KEY-----
+MIIFHzBJBgkqhkiG9w0BBQ0wPDAbBgkqhkiG9w0BBQwwDgQI9sGWjFBk5acCAggA
+MB0GCWCGSAFlAwQBKgQQnjti1hZOp7sXj+I6HDg1oQSCBNAkzqk+yAa6tYA3irhs
+f3Nw9CtC1oKO9Lgt6VVSyD1O9pmwzPKzi7C+0SHq4t57/d/wWEB5x5oggpgBui7D
+O+z3WTruSEqlJn3bit0g8Sx4rVsD8/vzGBUCVgOjmm6vUCUKwFOAK6Y6qRjL4c9W
+FB1RSGRn4jA++Ts76wstVJE+i6AXgkK9HSSVBStpjj3AAF8StQ9onGa6L+gUiw9J
+dCbEf9bP8hlY3tx1Ec+lp019VVxddufsydItPbXnDXuzSOW4AJ04KN1cenoh0ylW
+pKi8sJfRsN6/XWvDKWSxKc0inxVEeLXOeKMNue4+FSf2z+V5D8BtGdMFbwNhzDbP
+eIfG8myCfYiRZUeednB8ojG331j5H+TWfyGsG3POWZAH/VlCzne+4Yi/Nkjge7nd
+csZq7IV/UQMixsr0Paq0pRC26dYXLy2R8Z3gMffQ9sIFVEUbZ3L3DlqaGeXBZq34
+1+We9dTnUbIq3cL5fyhMl6hAlLcd8TD4xW+SKKYOmW0FrtjJdaoUax/w2tHA6dvk
+kbVGCBh1SfaT8/BaHNHAknLo9jE8SJPyTRuyLgHD+2oUlEIhIlBlJh9HvNYMjdnW
+eEGQzXSGDlTkjMT8LyAhqElX9EpOX/xyiLeAnZWq6pUgwIO+21rjaV0r+0ODwH7a
+EhAsx56++NfzQHf3Y2lSp5dRidhZlKVdEwvrsFaJ2bvlzY+VBKDVQs1YSytI//TG
+lYnRUQlGnlN3gAo1ZG0hixPBZixhblxNf7TB4MZ/Qu4rHk6SadceBO35lAbc7ltN
+N5moWRd6ayB6TVOXhl91FWCFPjyJFAn/pgH/u/1yBIBj35jOO3Hzn9itVQl4VWAI
+w/GyZyT0PRCRkTh+E5tL6Gzb8RyW0VQQwYOsoHeYHwO4L681fF8VJI4QbJ1UChrK
+Y37WzNwZXjemEGN41gj6IjZT4yGle0NLz7hIQJyIPTc+XxQ1O8CFiZJwYT/iIfqg
+hyumW/D9xEGlmNE8imgtWaoRmKnFds/cPFyEl5SxRXq6O8cNN5fgLABfwqbB/eFg
+ksrKhU0rL8t1LaxquUPFfc1Ae/A3d2hCgMrJaVdG/7Q9/WgsjhAisZHy7hvULT3+
+ejWWCEP3cd3iWrMli9wbHCF6mHSw9jYBqoWVc8K+BJwiZ+1BUHNT3Jm6kDSuLG9+
+b6TBfhACbGyqDNg1qNFcTy8KDtgaxktW5B3Y7unU8rsvNP7gt6pNrA7YWlcXQVWs
+E79WazfyMH/7pW7zb0eNu2rPp92/YxuZuSi9wCq9imymSIL+HAfmtARu29D3oIVU
+umFh37sIVxFSkf8a1WCTzVfK2uRIDGPIx8gr4ZzMwkV+87oZ3XRJKHIOuvZTHqUC
+p2pK7fqFurozkd/l1J41KExtxyECo+AGmh/zjU5Y9q+lXmPBPQ787PrHBYJHs7n/
++/vfdwI9SMItfElr4TZGLiVBHLhP15Bw8l4qGO0u4jcmKnb0WvyJi8H8O+L6893G
+s81jyKegHaV+mfyQAr7zrwhTN+MAcQxDJZgJS03MP0Z2usH4By+7Z7AFAzt/q5y2
+57Tl3M5dv8i3oidxqH7DN2JeOZpNOXSkksIw08QXuiF4Qs2ANKMEX/w6TTeU+K1v
+7wi1Tv+bHUmmqWaMnYhvxpcrrQ==
+-----END ENCRYPTED PRIVATE KEY-----
