--- src/main/resources/keystore/https/ca/arqonia_ca_certificate.pem
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDrzCCApegAwIBAgIJAJ0deQKkfFYZMA0GCSqGSIb3DQEBCwUAMG4xCzAJBgNV
-BAYTAlBMMRUwEwYDVQQIDAxXaWVsa29wb2xza2ExDzANBgNVBAcMBlBvem5hbjEQ
-MA4GA1UECgwHQXJxb25pYTEQMA4GA1UECwwHQXJxb25pYTETMBEGA1UEAwwKYXJx
-b25pYS5wbDAeFw0xOTAxMTcwOTQzMTVaFw0yOTAxMTQwOTQzMTVaMG4xCzAJBgNV
-BAYTAlBMMRUwEwYDVQQIDAxXaWVsa29wb2xza2ExDzANBgNVBAcMBlBvem5hbjEQ
-MA4GA1UECgwHQXJxb25pYTEQMA4GA1UECwwHQXJxb25pYTETMBEGA1UEAwwKYXJx
-b25pYS5wbDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJb8yzDsHkS7
-x8fedYls1UNM/ejXWvg2qAmIXTQw6/l2jlWQso/qql5Edd9TgXoe0sgWMSfCWsXT
-+De6gRyijl4zHUZ8zvbvjwQ/TFSqJRBvfxnlpQBG3HLIN2oSokL1ltvD354oZOig
-kEylr/EUDcTuG3pNaPFQyKLp149/v5A7ARcc/bRyyQThaQEIqtpdOnd83ESN4Rw3
-h2matTr/LLBKL1/3XPmb25Gc0l6YFoATwMl0Graeqn/SgaMcCUhBfiFi1GFxgudr
-tMavqSnqLXcfLDB/qzYLqNs6Pf0n/7va73lPHokH42AQRjas+5tbsoXru6TN84YK
-FCm8twSHNuMCAwEAAaNQME4wHQYDVR0OBBYEFEXUA+LJZR7rvL0lkPizXtB/Ziff
-MB8GA1UdIwQYMBaAFEXUA+LJZR7rvL0lkPizXtB/ZiffMAwGA1UdEwQFMAMBAf8w
-DQYJKoZIhvcNAQELBQADggEBAIi2hUpkYg9fkIz0R7m74gMRKwmyGfUSL+4afcT9
-xMvoa0tabq6mIj1/N2QvbK7/4jARjzFy/FHZTwxyCMII/rutLHQ40NVkj8tZrLuX
-t8eCq8ycX/nQupkSde2Pr0dtMGq8cnaeGTRo2sacrwkZfuHi49hmtjWA402xCdGR
-DdYPhdowRu8aq8ORvPxfkC3tEsYSppBo+H9EVvjO3mqSXrM7GbnqTE3FAf9X+hXP
-Uf/fk7pqoAj5jqDPvbrSJNUi1G6y5I/hFREgdQU+D2iRB6SwCtfo68wBCMm4EXno
-Y9HifdPUWL+5PuqO4922lMbgYkW6HPhZJJ9WEhHg75YvGHY=
------END CERTIFICATE-----
