--- src/main/resources/keystore/https/ca/arqonia_ca_certificate.srl
@@ -1,1 +1,0 @@
-8D9F4A7F17DF7350
