--- src/main/resources/keystore/https/ca/private/arqonia_ca_key.pem
@@ -1,30 +1,0 @@
------BEGIN ENCRYPTED PRIVATE KEY-----
-MIIFHzBJBgkqhkiG9w0BBQ0wPDAbBgkqhkiG9w0BBQwwDgQI7c5ybJ+xNJICAggA
-MB0GCWCGSAFlAwQBKgQQcpuz6LmFyBqaSbIVzc4dWgSCBNCQsCluqNtRfzN3S55w
-cqVjoNgPQlgIlg2mSPnc6KgFMpzCqmOVlG9sjNMQ5WdoOaIAkGtL2yfYSYBPxPv0
-WSWoiyOLN5Ewfr/kdYBJJvGsYIzQ2oiVzgipEjPxxPMC7MkaBmRa+qgJxMB0a1hB
-zogWivVr3tQ5BiXUIIKTpu6kg1+Bf4gPARDy9LvcrhkO64a4469iqbT4mFpvpxyW
-qOvLQV5hIoOXsQ0zFB/ekfKEEvpoh6UZ8l0SxMurZU8Q1L/7srzWJ4MiXHEoqvDB
-J5bRNJKF5sfYwjIGdLeIhtGh44yx8rFNaq29sQsOn5/8Ztj2FeWsbq3DNv5RNtZJ
-4M1lbsPga2/VdefFQgzfKiV1NRVpUp9ANjuA51MRNh1fltlAPba5WLKDGDAg32fl
-qrFS55MA6G8/tQ4BajSCoFfQ/1Ac2m0FbLjneYn5V7t4UanQ0EShVCt/AJJnfE9A
-kgORF/alBKhUc9WdLcd9yFsytDl2uBoTYS1JKKVfNLqCA2MBA9d7ZjhwxQm/xfXa
-VTBc0substCDNWE6TLFfrV/1UY0k7Z9Zi1psdHZkP1XI7jMyNMOk454liauGKmrj
-DjDL3qabdDtxHLEZD/8aJoPJL6AVIDg+H3NLVx+act47LRT3vgS3rmSDTQc0dUHv
-5IIIKyKQD51rp0kIEGBf87cShxOGC/Wci7rctaLEyCwYz8C56oddOEW2qr7eUEv+
-NCJV34rX7mtmb17hA7fZTCb4wh5r35PCmYRdOqaf8miouThp8i8pior60E/KuhGG
-ZYQ7nMABtqYYkZ5opmGDVORkh1bQ6DfuoOyTE/qHdxOAKJzCui+MC8Ghdc33iljV
-UbRkSnhPoft4DZhsjLIPaZNFTcRfw84wvKjZctvfFy23I2we8Vl9FfviG917IKma
-LlDTC9GPJWWvSa9tT4JdI2YUHNTExFD7opk/qSBERLAceKtmuUYOSoib03J3rAsi
-i2/IAhCB4ud6oHOfa5XgRpx2uUZ8Pv1+HWpF0bDkehOjkHlm1PucgcJjULuV4UjO
-kd3YoEfRyS0nfTr5uwsyKu9AswBlbMVwVXRDbsgVPHPwjE9HRwVOvd2CUOy+5nS2
-+6myqawo6/C8wpG5/C75gnUX/y959gTzapMH8UShDKLKdwNNMDo9+V9yjXnRVCxk
-2+R8aC5dAnUWVIbzAqXYI0JgawwD93wszj87qS0k+xs9O0A1NPfHzaDZwb3BqmtM
-vbsFrn0NpV7faOLb7aAtyQGOEzocgFiA+LBbhHh1aUZRffM0bYhocPdGZmYIKcry
-/2I6urZlNZhZzJ4qVKfPCfbqsxuM+ZWcZG2VgBlR4EBuU9QxitzJPwsj5+cUqsPd
-jVGEHuyYCkkaX0e0bfz7TESBF7LdcMqfKJLNKzmXpQN7axcoC38KIK6JkVxA2b/h
-3xF7c1YW2AC11JewxMo7izKMhhjxOzXycqR1caeeNAhjtNqlISmGkz87E03UGPcP
-ir5ZGYK8tqQXsOz40F9MQ1EISCaDj2F4z9zf/awFYS3lnuJjpFXWjxtaeTQ1OIFi
-wkeQRURW44fUIgZZMfu9Q0Tfs+2pkZvJGaujTq0lldW0Yw2ry0Gp+aDN+WqrIAf3
-gsxuMf3fL8pRg4OJHnaDv2AMUQ==
------END ENCRYPTED PRIVATE KEY-----
